Create CSharp_Out_Keyword_Practice.cs
Discovered the "out" keyword in C#. Wrote a small program to test it out.

The purpose is to use a variable that's not initialized in the first place, and have the variable declaration and operations be handled by the method. Found out about this when implementing a Queue in another program.