Create QueuePractice.cs
Learning about the Queues data structure. Took some example code, and deconstructed it. Also learned about C#'s "out" keyword, which is useful for when I don't want to initialize a variable. In this case: useful because I'm using a Generic parameter (could be a LinkedList of Objects, Ints, Strings, etc.)